feat: add all sample routes
