+++ app/(tools)/dump-emmc/page.tsx
@@ -1,0 +1,3 @@
+export default function SampleRoute() {
+  return <div>Sample Route, to be populated later.</div>
+}
